test: 네트워크 요청을 가짜 응답으로 대체할 수 있는 MockURLProtocol 구현
- URLSession.protocolClasses를 활용해 MockURLProtocol 주입 가능
- 실제 네트워크 의존성 제거
- APIService 정상/에러 응답 시나리오 테스트 가능
